Decrease memory usage (+ other changes)
Make display lists dynamically sized! All chunks are not the same and
should not be treated as such.
This fixes a bug of accidentally overrunning the fixed sized
Change the default size of a display list to 256.

Demote various blocks to render pass 0, since zcomploc was fixed only
water should be in pass 1
Add world seeding! Each run will give you a new world.
Display the seed in the top right corner (hexadecimal)
Make the player timer a double and fix the slowdown at lower FPS
Dim the inventory selector if in flying mode.
Add lowercase letters to the font and lowercase various text
Add more precision to the memory usage display

+++ source/player.h
@@ -12,5 +12,5 @@ typedef struct player_s {
 	u8 inventory[10];
 	bool flying;
 	bool select;
-	unsigned short timer;
+	double timer;
 } player;
